EE.jl API

The EE.jl API dynamically wraps the EarthEngine Python API, so it defines the types and methods on-the-fly when initializing a session. This means there are few actual Julia functions defined mostly meant to create Julia methods and types from the Python API (see below).

For more in depth documentation on specific methods for using the EarthEngine API, see the official Earth Engine Documention.

EE.Initialize — Method

Initialize(args...; kwargs...)

Function to initialize an EarthEngine session (analagous to ee.Initialize() from the Python API). Accepts arguments and keywords from the Python ee.Initialize() function. This function also dynamically builds the EE API and creates the methods with signatures for each EE Type.

EE.Authenticate — Method

Authenticate()

Function to execute the EarthEngine authetication workflow (analgous to ee.Authenticate() in the Python API). This function should only be executed once if the EE API has not be used before.

EE.ee_wrap — Function

ee_wrap(pyo::PyObject)

Function for wrapping a Python object defined in the type map

EE.pyattr — Function

pyattr(class method orig_method)

Function for creating a method signature for a Julia Type In Python world method(T::class) is analagous to class.method()

EE.@pytype — Macro

@pytype name class

Macro for creating a Julia Type that wraps a PyObject class

EE.pyattr_set — Function

pyattr_set(types, methods...)

For each Julia type T<:EEWrapped in types and each method m in methods, define a new function m(t::T, args...) that delegates to the underlying pyobject wrapped by t.

EE.@pyattr — Macro

@pyattr class method

Macro for creating a method signature for a Julia Type In Python world method(T::class) is analagous to class.method()

@pyattr class method orig_method

Macro for creating a method signature for a Julia Type In Python world method(T::class) is analagous to class.method() This will create a new method name which calls the orig_method