EarthEngine.jl API

The EarthEngine.jl API dynamically wraps the EarthEngine Python API, so it defines the types and methods on-the-fly when initializing a session. This means there are few actual Julia functions defined mostly meant to create Julia methods and types from the Python API (see below).

For more in depth documentation on specific methods for using the EarthEngine API, see the official Earth Engine Documention.

EarthEngine.Initialize — Method

Initialize(args...; kwargs...)

Function to initialize an EarthEngine session (analagous to ee.Initialize() from the Python API). Accepts arguments and keywords from the Python ee.Initialize() function. This function also dynamically builds the EE API and creates the methods with signatures for each EE Type.

EarthEngine.@eefunc — Macro

@eefunc method

Macro to wrap Julia defined method on the Python side. This is needed when applying functions to specific types such as EE.List.

@eefunc method type

Macro to wrap Julia defined method with a specific type signature on the Python side. This is needed when applying functions with type annotations using map()

EarthEngine.eefunc — Function

eefunc(method)

Function to wrap Julia defined method on the Python side. This is needed when applying functions to specific types such as EE.List.

EarthEngine.ee_wrap — Function

ee_wrap(pyo::PyObject)

Function for wrapping a Python object defined in the type map

EarthEngine.pyattr — Function

pyattr(class method orig_method)

Function for creating a method signature for a Julia Type In Python world method(T::class) is analagous to class.method()

EarthEngine.@pytype — Macro

@pytype name class

Macro for creating a Julia Type that wraps a PyObject class

EarthEngine.pyattr_set — Function

pyattr_set(types, methods...)

For each Julia type T<:AbstractEEObject in types and each method m in methods, define a new function m(t::T, args...) that delegates to the underlying pyobject wrapped by t.

EarthEngine.@pyattr — Macro

@pyattr class method

Macro for creating a method signature for a Julia Type In Python world method(T::class) is analagous to class.method()

@pyattr class method orig_method

Macro for creating a method signature for a Julia Type In Python world method(T::class) is analagous to class.method() This will create a new method name which calls the orig_method